requirements.prod.txt
discord-py-interactions~=5.12
requests~=2.31
urllib3~=2.2
pyyaml~=6.0